remove entire old navigation
