+++ features/ajna/common/layout.tsx
@@ -3,7 +3,7 @@ import { WithFeatureToggleRedirect } from 'components/FeatureToggleRedirect'
 import { PageSEOTags } from 'components/HeadTags'
 import { BasicLayout } from 'components/Layouts'
 import { AnjaFooter } from 'features/ajna/common/components/AnjaFooter'
-import { AjnaNavigationController } from 'features/ajna/common/controls/AjnaNavigationController'
+import { AjnaNavigationController } from 'features/navigation/controls/AjnaNavigationController'
 import React, { PropsWithChildren } from 'react'
 import { ajnaExtensionTheme } from 'theme'
 import { ThemeProvider } from 'theme-ui'
